fix security and stability issues

+++ migrations/2026_01_10_000004_add_roleplay_experience_to_character_sheets.php
@@ -11,7 +11,9 @@ return [
 
         if (!$schema->hasColumn('character_sheets', 'roleplay_experience')) {
             $schema->table('character_sheets', function (Blueprint $table) {
-                $table->unsignedTinyInteger('roleplay_experience')->default(0)->after('sum');
+                // Clamped 0-999 in CharacterSheetParser::extractRoleplayExperience,
+                // so this needs more range than TINYINT UNSIGNED (max 255) gives.
+                $table->unsignedSmallInteger('roleplay_experience')->default(0)->after('sum');
                 $table->index(['roleplay_experience'], 'character_sheets_rpexp_idx');
             });
         }
